Refactor: rename test_hit.py to test_suggestion.py, fix method and var names. Part of #267
